Changed cplusplus/using_declarations
Changed to use size_t instead of int.

+++ cplusplus/using_declarations/with_explicit_qualification.cpp
@@ -15,13 +15,13 @@ static int solve(const std::string &a,const std::string &b)
     //  L01111
     //  Y01112
 
-    int n = a.size();
-    int m = b.size();
+    size_t n = a.size();
+    size_t m = b.size();
 
     std::vector<std::vector<int>> table(n+1,std::vector<int>(m+1,0));
 
-    for (int i=0; i!=n; ++i) {
-        for (int j=0; j!=m; ++j) {
+    for (size_t i=0; i!=n; ++i) {
+        for (size_t j=0; j!=m; ++j) {
             if (a[i]==b[j]) {
                 table[i+1][j+1] = table[i][j] + 1;
             }
